engine: add ability to have non-dbus sources
Add the (theoretical) ability to have a non-writable source that is not
associated with a D-Bus instance by adding checks to ensure the bus type
is not equal to G_BUS_TYPE_NONE before performing D-Bus match rule
operations from the engine.

We don't perform checks before any other D-Bus operations because the
only other operation is a write and we only expect to support
non-writable sources in this manner.