engine: add support for runtime profile selection
Add support to dconf-engine for opening "runtime" profiles.

These profiles are intended to be symbolic links or plain files that
will live either in XDG_RUNTIME_DIR/dconf/profile or
/run/dconf/user/$(uid).

This is intended to allow for a PAM module that makes complex decisions
about application of a specific policy to a user and sets up the profile
at login time, thus preventing the need for this complex decision to be
a part of every program that uses dconf.  This PAM module would not be
part of dconf, but would rather be a part of a dconf-aware system
administrator framework.

In the case that the profile file is found in /run/dconf, then it will
not be possible for the user to override the profile selection,
including via the DCONF_PROFILE environment variable.  This provides a
mechanism for lockdown that is slightly more difficult for a user to
circumvent.  In theory, this is pointless since it can still be defeated
with LD_PRELOAD, but in practice this raises the bar quite a bit.

https://bugzilla.gnome.org/show_bug.cgi?id=751417