writer: add a signal to notify for errors when writing
Some clients will be interested to know when a write failed for
specific reasons, e.g. when there's no space left on the device.

This commit adds a signal to that effect. For now we differentiate
between two error codes: a generic one and one for when the writer is
out of disk space.

Fixes: #29