gvdb: remove memory allocation from GvdbPath
Modify GvdbPath to never allocate memory.

Previously, if we had fewer split points ('/') than the number of
pre-allocated items in the arrays in the structure, we would simply use
them.  At 16, this number is already extremely high, and it's
implausible to imagine a real case for which this would be insufficient.

This commit simplifies things a bit: if there are more than 16 segments,
we will just ignore the later ones, except for the final one (ie: the
complete path).

For the sake of an example, let the limit be 4, rather than 16.  This
means that you could lock:

  /
  /org/
  /org/gnome/
  /org/gnome/app/deeply/nested/key

but not:

  /org/gnome/app/
  /org/gnome/app/deeply/
  /org/gnome/app/deeply/nested/

With 16 segments, everything here could be locked, and much more.

In this way, we preserve the previous behaviour of always being able to
lock a particular individual key of any depth, while introducing
path-based locks for all reasonable cases, and we avoid memory
allocations in all cases.