engine: rework initial value for has_locks
In the case that the user database file does not exist on disk,
"re"opening it for the first time will fail, causing the refresh
function to return FALSE.  This means that we will not end up
recomputing the value of has_locks as was previously assumed.

To avoid this problem, we fill in the proper value from the start,
instead of guessing.