engine: use GvdbPath for writability checks
Move over to using GvdbPath when checking for writability of a key.

This has two advantages:

The first is that we only hash the key once during writability checks,
even if we have multiple stacked databases.

The second is that we can now lock down entire subpaths in dconf.

The way the code is written also means that it is now theoretically
possible to "unlock" a given path or key, which means that a database
can introduce a lock for "/" but unlock "/org/gnome/myapp/", in effect,
preventing writes to any area outside of that path.  The "best" (ie:
most specific) result is taken as authorative.  These 'negative locks'
are not (yet?) supported in the dconf(1) update/compile commands, but
they will be used for proxied databases for application confinement.

Note: each database is consulted separately.  That means that a
higher-level database cannot undo a lock of a lower-level database with
a more-specific unlock.  The security model is therefore the same as
what it was before.