engine: remove spurious local change notifications
When used with the "fast" (optimistic concurrency) API, the engine library
emits a change notification local to a process after a change is initiated
from that process. Previously, it would emit this notification even if the
newly written value was the same as the previous value (according to that
process's view of the state). After this change, the local change
notification is not sent unless the new value is different from the
current value (as seen by that process).