build: Use completions dir from pkg-config file
The `bash-completions` pkg-config file provides a variable with the
location of the completion files.

This variable is checked to set the installation directory of the
completion file provided by dconf.

As a side note, the variable in the `bash-completions` pkg-config
file is not totally correct, because it provides a directory
that is relative to prefix while it should be relative to datadir.

https://bugzilla.gnome.org/show_bug.cgi?id=790236