Use the correct Meson argument for linking
Some of the internal static libraries in dconf implement public symbols.
This means that they must be linked using `link_whole`, to prevent the
linker from hiding unused symbols in there when building the shared
libdconf.so.

This fixes the build of projects using the dconf API directly, like
dconf-editor, in GNOME Continuous.

Reviewed-by: nobody
Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>