Namespace the dependencies variables
The build currently, unintentionally, depends on side effects, namely:
that a variable defined in a meson.build file will be valid for any
other included meson.build file, until it gets re-defined.

We use the same `deps` variable in many places, and we end up depending
on the inclusion order when we get to the client library, which depends
on the gsettings backend defining the `deps` variable.

Reviewed-by: nobody
Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>