Avoid use of link_whole in the gsettings backend
The GSettings backend loadable module should not expose any symbol
except for the g_io_module_* ones. By using the internal dependency
objects, we end up using `link_whole` on libdconf-common and
libdconf-gdbus-thread, which means that we now inherit all the symbols
exposed by those static libraries.

We still want to use `link_whole` every time we depend on those static
libraries, but for the GSettings backend module we make an exception,
and use `link_with` with the build targets, instead of the internal
dependencies.

This fixes the `abicheck` test unit.

Reviewed-by: nobody
Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>