build: Remove libdconf-common-hidden
`libdconf-common-hidden` is built to hide symbols not necessary
in the `gsettings` gio module that only needs to expose
`g_io_module_[load|query|unload]` symbols.

To achieve this a symbol map along with `version-script` linker flag
is used. Thanks to this only the necessary symbols are exposed,
building `libdconf-common-hidden` static library is not necessary
anymore and the existing dependencies can be used.