build: Use weak bindings in gvdb to fix linking with LLD
Since tests/dconf-mock-gvdb.c has functions conflicting with the real
gvdb and it is intended for the former to override the latter in tests,
so we have to make functions in gvdb library have weak bindings instead
of the default strong bindings to avoid duplicate symbol errors.

This change should not affect exported symbols of shared libraries
because these symbols are already have visibility set to hidden.

Fixes https://gitlab.gnome.org/GNOME/dconf/issues/47