tests: shm: fix pwrite wrapper with -D_FILE_OFFSET_BITS=64
Due to the hacks used for large file support, wrapping pwrite is error
prone and can end up calling the wrong function. Currently, "pwrite" is
called instead of "pwrite64" on 32-bit ARM, causing the test to fail.
This commit attempts to determine the correct symbol to call from the
wrapper.