gdbus/thread: Always copy GError on getters
This change should guarantee that the object is valid while in use.

The change is part of a series of changes to fix an issue
introduced with recent changes made to GLib[1] where invoking
g_test_dbus_down() will fail after a timeout if the GDBusConnection
object for the session bus leaks.

[1] https://gitlab.gnome.org/GNOME/glib/merge_requests/963

Signed-off-by: Andre Moreira Magalhaes <andre@endlessm.com>