build: Update use of link_whole for meson-0.52
A regression in meson-0.52 caused uses of link_whole to expose scenarios
where duplicate symbols issues could appear. In particular
libdconf_client_dep was being link_whole'd to itself, which recursively
already included libdconf_common which was also a link_whole.

This change does not modify the available symbols in libdconf.so, and is
compatible with meson-0.52 and 0.51.

See: https://github.com/mesonbuild/meson/pull/6030
Fixes: https://gitlab.gnome.org/GNOME/dconf/issues/59