gsettings: Update abicheck.sh script
gcc-9 is including `mangle_path` in the symbols of
gsettings/libdconfsettings.so when building with support for gcov
(--coverage). This means that our ignored symbols had to be updated.

Update abicheck.sh so the test suite passes again. Also include a
comment to explain the above and how the test works.

Fixes: https://gitlab.gnome.org/GNOME/dconf/issues/60