bin/completion/dconf: Add compile to autocomplete
While working with Dconf I noticed in the bash CLI
I was unable to autocomplete when typing "dconf compile"

This fix will make it so the user is able to auto complete
the compile option in bash.
I have tested it on my Manjaro and PopOS! machines
and it's working expected.

As this is my first contribution please look twice so I have
not missed anything.

Signed-off-by: Andreas Polnas <andreas.polnas93@hotmail.com>